+++ nginx-auth/lua/config.lua
@@ -15,11 +15,17 @@ config = {
         mode = 'login-form',
 
         -- secret used for signing cookies with hmac
-        hmac_secret = 'franz jagt im komplett verwahrlosten taxi quer durch bayern.',
+        hmac_secret = 'franz jagt im komplett verwahrlosten taxi quer durch bayern',
 
         -- cookie expiration time
-        --cookie_expiration = 60 * 60,    -- 1 hour
-        cookie_expiration = 1 * 60,     -- 1 minute
+        cookie_expiration   = 60 * 60 * 3,  -- 3 hours
+        --cookie_expiration = 60 * 60,      -- 1 hour
+        --cookie_expiration = 1 * 60,       -- 1 minute
+        --cookie_expiration = 10,           -- 10 seconds
+
+        -- cookie renewal threshold
+        -- if expiration time delta is below this threshold, automatically issue a new cookie
+        cookie_renewal      = 60 * 60,      -- 1 hour
 
         -- name of the cookie
         cookie_name = 'Auth',