+++ elmyra.ip.access.epo/CHANGES.rst
@@ -11,6 +11,7 @@ development
 - middleware/auth: nginx-lua-auth proof-of-concept (http basic auth)
 - middleware/ui/auth: nginx-lua-auth enhancements (login-form)
 - ui/auth: enhance login box: integrate actions (login failed, register account) via email; rumble on error
+- auth: automatic cookie renewal, proper error responses for /api and /static routes
 
 0.29.0
 ======